factoring plot tools
